erasure_code: constify a bunch of "src" args
The reasoning for this commit is to constify ec_encode_data() function.
It is expected that after this function has run, the data that "src"
argument points to has not changed. Hence this argument should really be
`const`.

All other functions that undergo same changes here are simply called by
ec_encode_data() internally, one way or another.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>